refactor(backend): split triggers & sessions repositories into contract/impl
Restructure both DynamoDB repositories as deep modules with a narrow public
surface, per the deep-modules skill:

- triggers/ and sessions/ now expose a TriggersRepository / SessionsRepository
  interface (the behaviour contract) plus a domain model in types.ts; the
  DynamoDB implementation lives under dynamodb/ and is reached only by the
  composition-root factory.
- Domain types carry no storage keys (PK/SK/GSI). Storage<->domain mapping is
  isolated in dynamodb/item.ts and projects onto a domain-field allowlist
  (pick) instead of stripping a storage-key denylist, so a new index key can
  never leak.
- Method input/output types co-locate with their interface; the barrel is the
  single entry point and re-exports the contract explicitly.
- Add libs/object/pick: a type-checked projection helper shared by the mappers.
- getExecutions now strips storage keys (was inconsistent with the other reads)
  and backfills the legacy startedAt timestamp in the repository, removing the
  route-level back-compat hack.

Behaviour-preserving: integration suites (26) and unit suites (106) pass.
Also extends the deep-modules SKILL.md with the allowlist-mapper, deliberate-
interface-extraction, and split-to-narrow-surface guidance this work surfaced.
